#29: Move getUserInput method to CommandLineInputService.
