--- src/main/java/org/weathertrack/input/UserInputService.java
@@ -1,5 +1,0 @@
-package org.weathertrack.input;
-
-public interface UserInputService {
-	String getUserInput(String message);
-}
