+++ src/main/java/org/weathertrack/input/service/InputService.java
@@ -1,10 +1,12 @@
-package org.weathertrack.input;
+package org.weathertrack.input.service;
 
 import org.weathertrack.weather.model.WeatherData;
 
 import java.util.List;
 
-public interface UserInputInterface {
+public interface InputService {
+	String getUserInput(String message);
+
 	String getCityNameFromUser();
 
 	void printCitiesWithSameName(List<String> citiesWithSameName);
